# backend/docs/swagger.yaml
definitions:
  schemas.Action:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      option:
        items:
          type: integer
        type: array
      service:
        $ref: '#/definitions/schemas.Service'
      update_at:
        type: string
    required:
    - description
    - name
    - option
    - service
    type: object
  schemas.Area:
    properties:
      action:
        $ref: '#/definitions/schemas.Action'
      action_option:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      description:
        type: string
      enable:
        type: boolean
      id:
        type: integer
      reaction:
        $ref: '#/definitions/schemas.Reaction'
      reaction_option:
        items:
          type: integer
        type: array
      title:
        type: string
      update_at:
        type: string
      user:
        $ref: '#/definitions/schemas.User'
    required:
    - action
    - action_option
    - description
    - reaction
    - reaction_option
    - title
    - user
    type: object
  schemas.AreaMessage:
    properties:
      action_id:
        description: Foreign key for Action
        type: integer
      action_option:
        items:
          type: integer
        type: array
      description:
        type: string
      reaction_id:
        description: Foreign key for Reaction
        type: integer
      reaction_option:
        items:
          type: integer
        type: array
      title:
        type: string
    required:
    - action_option
    - description
    - reaction_option
    - title
    type: object
  schemas.AuthenticationURL:
    properties:
      authentication_url:
        type: string
    type: object
  schemas.CodeCredentials:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  schemas.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  schemas.JWT:
    properties:
      token:
        type: string
    type: object
  schemas.Reaction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      option:
        items:
          type: integer
        type: array
      service:
        $ref: '#/definitions/schemas.Service'
      update_at:
        type: string
    required:
    - description
    - name
    - option
    - service
    type: object
  schemas.Response:
    properties:
      message:
        type: string
    type: object
  schemas.Service:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        $ref: '#/definitions/schemas.ServiceName'
      oauth:
        type: boolean
      update_at:
        type: string
    required:
    - color
    - description
    - icon
    - name
    - oauth
    type: object
  schemas.ServiceName:
    enum:
    - Spotify
    - OpenWeatherMap
    - Timer
    - Gmail
    - Github
    - Dropbox
    - Microsoft
    type: string
    x-enum-varnames:
    - Spotify
    - Openweathermap
    - Timer
    - Gmail
    - Github
    - Dropbox
    - Microsoft
...
  schemas.User:
    properties:
      created_at:
        description: Time when the user was created
        type: string
      email:
        description: Email of the user
        type: string
      id:
        description: Unique identifier for the user
        type: integer
      password:
        description: can be null for Oauth2.0 users
        type: string
      token_id:
        description: Foreign key for LinkURL
        type: integer
      updated_at:
        description: Time when the user was last updated
        type: string
      username:
        description: Username of the user
        type: string
    required:
    - email
    - username
    type: object
...
  schemas.UserCredentials:
    properties:
      email:
        description: Email of the user
        type: string
      username:
        description: Username of the user
        type: string
    type: object
info:
  contact: {}
paths:
...
  /area:
    delete:
      consumes:
      - application/json
      description: delete user area list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Area'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - Bearer: []
      - bearerAuth: []
      summary: delete user area
      tags:
      - Area
    get:
      consumes:
      - application/json
      description: get user areas list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.Area'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - Bearer: []
      - bearerAuth: []
      summary: get user areas
      tags:
      - Area
    post:
      consumes:
      - application/json
      description: create area
      parameters:
      - description: Area Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.AreaMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: create area
      tags:
      - Area
    put:
      consumes:
      - application/json
      description: update user area list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Area'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - Bearer: []
      - bearerAuth: []
      summary: update user area
      tags:
      - Area
  /dropbox/auth:
    get:
      consumes:
      - application/json
      description: give url to authenticate with dropbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AuthenticationURL'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with dropbox
      tags:
      - Dropbox
  /dropbox/auth/callback:
    post:
      consumes:
      - application/json
      description: give url to authenticate with dropbox
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with dropbox
      tags:
      - Dropbox
  /dropbox/auth/callback/mobile:
    post:
      consumes:
      - application/json
      description: give url to authenticate with dropbox
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give url to authenticate with dropbox
      tags:
      - Dropbox
  /dropbox/file:
    get:
      consumes:
      - application/json
      description: give user info of dropbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of dropbox
      tags:
      - Dropbox
  /dropbox/folder:
    get:
      consumes:
      - application/json
      description: give user info of dropbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of dropbox
      tags:
      - Dropbox
  /dropbox/info:
    get:
      consumes:
      - application/json
      description: give user info of dropbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserCredentials'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of dropbox
      tags:
      - Dropbox
...
  /gmail/auth:
    get:
      consumes:
      - application/json
      description: give url to authenticate with gmail
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AuthenticationURL'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with gmail
      tags:
      - Gmail
  /gmail/auth/callback:
    post:
      consumes:
      - application/json
      description: give url to authenticate with gmail
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with gmail
      tags:
      - Gmail
  /gmail/auth/callback/mobile:
    post:
      consumes:
      - application/json
      description: give authentication token to mobile
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give authentication token to mobile
      tags:
      - Gmail
  /gmail/info:
    get:
      consumes:
      - application/json
      description: give user info of gmail
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserCredentials'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of gmail
      tags:
      - Gmail
...
securityDefinitions:
  bearerAuth:
    description: Use "Bearer <token>" as the format for the Authorization header.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"